+++ frontend/lib/pages/settings/settings_page.dart
@@ -1,8 +1,9 @@
 import 'dart:developer';
 import 'package:flutter/material.dart';
-import 'package:frontend/pages/all_pages_export.dart';
 import 'package:settings_ui/settings_ui.dart';
 import 'package:frontend/utils/all_utils.dart';
+
+import '../../widgets/speaq_settings_section.dart';
 
 class SettingsPage extends StatefulWidget {
   const SettingsPage({Key? key}) : super(key: key);
@@ -106,37 +107,4 @@ class _SettingsPageState extends State<SettingsPage> {
   }
 }
 
-class SpqSettingsSection extends AbstractSettingsSection {
-  const SpqSettingsSection({
-    required this.tiles,
-    this.margin,
-    this.title,
-    Key? key,
-  }) : super(key: key);
 
-  final List<Widget> tiles;
-  final EdgeInsetsDirectional? margin;
-  final Widget? title;
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      mainAxisAlignment: MainAxisAlignment.start,
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        Container(
-            padding:
-                const EdgeInsets.symmetric(vertical: 2.0, horizontal: 24.0),
-            child: title),
-        Container(
-          child: Column(
-            children: tiles,
-          ),
-          decoration: const BoxDecoration(
-              color: spqWhite,
-              borderRadius: BorderRadius.all(Radius.circular(24.0))),
-        )
-      ],
-    );
-  }
-}
